Fix a subtle bug introduced by refactoring merges that didn't come up until level 11. If as a result of the move two jellies begin touching in more than one place, the first pair of touching cells will unify the jellies and the second won't even call mergeWith, so the borders will remain.
A bit ugly, but the only simple way I see of fixing is to allow merging
two cells within the same jelly, for proper handling of borders. To prevent
a huge number of null DOM updates, track directions in which each jellycell
already merged, and don't even try merging there.